Bug 826613 - Enable FAIL_ON_WARNINGS on MSVC in image/. r=joe
UltraBlame original commit: 80418f0fc73d9617499f9e9856420fd27a2771f8

+++ image/src/imgRequest.cpp
@@ -767,7 +767,8 @@ imgRequest::OnDataAvailable(nsIRequest *aRequest, nsISupports *ctxt,
       
       
       mImage = ImageFactory::CreateImage(aRequest, mStatusTracker, mContentType,
-                                         mURI, mIsMultiPartChannel, mInnerWindowId);
+                                         mURI, mIsMultiPartChannel,
+                                         static_cast<uint32_t>(mInnerWindowId));
 
       
       mStatusTracker = nullptr;
